Created a rough tilemap with placeholder assets
Assets are placeholders I found online, also created a folder to store the assets for the levels and tilemaps. The Walls are not hitboxed perfectly would like to change once we have real assets for the tiles. PLAYER WAS NOT ABLE TO MOVE FOR ME. ERROR: COULD NOT ACCESS PLAYER SCRIPT.